// packages/nx/src/schematics/application/schema.d.ts
export interface Schema {
  name: string;
  framework?: string;
  routing?: boolean;
  skipFormat: boolean;
  directory?: string;
  tags?: string;
  unitTestRunner: 'jest' | 'none';
  /**
   * Group by app name (appname-platform) instead of the default (platform-appname)
   */
  groupByName?: boolean;
}